Consolidate exacqman and exacqvision CLIs for cleaner homebrew distro
